--- pydapsys/rawio/__init__.py
@@ -1,6 +1,0 @@
-from __future__ import annotations
-
-from typing import Literal
-
-INT_STRUCTS = Literal['b', 'B', 'h', 'H', 'i', 'I', 'l', 'L', 'n', 'N', 'P']
-FLOAT_STRUCTS = Literal['e', 'f', 'd']
